Migrate account_move_locking to 10.0
