+++ account_move_locking/README.rst
@@ -19,7 +19,7 @@ In order to lock the journal entry, you need to follow this process:
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/92/9.0
+   :target: https://runbot.odoo-community.org/runbot/92/10.0
 
 Bug Tracker
 ===========
@@ -35,7 +35,8 @@ Credits
 Contributors
 ------------
 
-* Vincent Renaville <vincen.renaville@camptocamp.com>
+* Vincent Renaville <vincent.renaville@camptocamp.com>
+* Akim Juillerat <akim.juillerat@camptocamp.com>
 
 Maintainer
 ----------
