+++ account_move_locking/views/account_view.xml
@@ -1,4 +1,4 @@
-<openerp>
+<odoo>
   <data>
 
     <record id="view_move_form_locked" model="ir.ui.view">
@@ -13,4 +13,4 @@
     </record>
 
   </data>
-</openerp>
+</odoo>
