+++ account_move_locking/wizard/account_lock_account_move.py
@@ -1,27 +1,9 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author Vincent Renaville.
-#    Copyright 2015 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-##############################################################################
 
-from openerp import fields, models, api, _, exceptions
+from odoo import fields, models, api, _, exceptions
 
 
-class lock_account_move(models.TransientModel):
+class LockAccountMove(models.TransientModel):
     _name = "lock.account.move"
     _description = "Lock Account Move"
