+++ account_move_locking/wizard/account_lock_move_view.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
+<odoo>
     <data>
 
         <!--Account Moves-->
@@ -15,7 +15,6 @@
                     </group>
                     <footer>
                         <button string="Approve" name="lock_move" type="object" class="oe_highlight"/>
-                        or
                         <button string="Cancel" class="oe_link" special="cancel"/>
                     </footer>
                 </form>
@@ -39,4 +38,4 @@
            />
 
     </data>
-</openerp>
+</odoo>
